+++ Modules/Accounts/Components/LegalAgreement/LegalAgreement.cs
@@ -36,7 +36,7 @@ namespace Unity.AI.Toolkit.Accounts.Components
             packages = new()
             {
                 "com.unity.ai.generators",
-                "com.unity.ai.chat"
+                "com.unity.ai.assistant"
             },
 
             installButtonText = "Agree and install AI features"