+++ Modules/Accounts/Components/SessionStatusBanner/Content/AIDisabledBanner.cs
@@ -8,7 +8,7 @@ namespace Unity.AI.Toolkit.Accounts.Components
     public partial class AIDisabledBanner : BasicBannerContent
     {
         public AIDisabledBanner() : base(
-            "AI features are not enabled. Your administrator has disabled AI features for your organization. <link=manageaccount><color=#7BAEFA>Manage account</color></link>",
+            "Your organization has disabled the use of Unity AI. Contact your organization’s managers to use these packages. <link=manageaccount><color=#7BAEFA>Manage account</color></link>.",
             new LabelLink("manageaccount", AccountLinks.ManageAccount)
         ) { }
     }