+++ Modules/Accounts/Components/SessionStatusBanner/Content/AIDisabledLegalBanner.cs
@@ -1,0 +1,11 @@
+using System;
+using UnityEngine.UIElements;
+
+namespace Unity.AI.Toolkit.Accounts.Components
+{
+    [UxmlElement]
+    public partial class AIDisabledLegalBanner : BasicBannerContent
+    {
+        public AIDisabledLegalBanner() : base("Legal agreement is missing.") { }
+    }
+}