+++ CHANGELOG.md
@@ -4,18 +4,25 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-pre.4] - 2025-04-16
+
+### Changed
+
+- Update SDK to version 0.16.2.
+- Update assistant plugins.
+- Update ai-enabled for assistant and generators.
+
 ## [1.0.0-pre.3] - 2025-04-09
 
-### Fixed
+### Changed
 
-- Update SDK to version 0.15.0
+- Update SDK to version 0.15.0.
 
 ## [1.0.0-pre.2] - 2025-03-27
 
 ### Changed
 
 - Moved shared modules to generators namespace.
-
 
 ## [1.0.0-pre.1] - 2025-03-21